pyproject.toml
[tool.poetry]
name = "csv-read"
version = "0.1.0"
description = ""
authors = ["Bolot Zhusupekov <bolot.zhusupekov07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
pandas = "^2.2.2"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
django-environ = "^0.11.2"
celery = "^5.4.0"
psycopg2 = "^2.9.9"
django-debug-toolbar = "^4.4.5"
requests = "^2.32.3"
django-redis = "^5.4.0"
black = "^24.4.2"
django-filter = "^24.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
